pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["atium"]

[project]
name = "atium"
version = "0.0.1"
requires-python = ">=3.12"
authors = [{ name = "Shrenik", email = "shrenik95@gmail.com" }]
description = ""
readme = "README.md"
license = { file = "LICENSE" }
classifiers = []
dependencies = [
  # Core
  "attrs==25.3.0",
  "pytest==7.4.3",
  "pyyaml==6.0.2",
  "nptyping==2.5.0",

  # Visualization/tooling
  "matplotlib==3.10.1",
  "opencv-python==4.11.0.86",
  "notebook==7.4.0",

  # Numerical
  "numpy==1.26.4",
  "ortools==9.12.4544",
  # TODO: Separate install for CPU and GPU versions.
  "jax==0.5.3",
  "osqp==1.0.3",
  "drake==1.40.0",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
py_version = 312
skip_gitignore = "True"
multi_line_output = 3
line_length = 120

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"
respect-gitignore = true


[tool.ruff.lint]
select = [
  # Pep8
  "E",
  # Pyflakes
  "F",
  # Isort
  "I",
  # Bugbear
  "B",
  # Annotations
  "A",
  # For black compatible formatting
  "RUF100",
  # Remove unused imports
  "F401",
]
# Fix re-organize imports and unused imports
# These will provide code actions.
fixable = ["I001", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
